+++ packages/terra-date-picker/tests/wdio/date-picker-spec.js
@@ -328,4 +328,21 @@ describe('Date Picker', () => {
     Terra.should.matchScreenshot();
     Terra.should.beAccessible({ rules: ignoredA11y });
   });
+
+  describe('Key Limitations', () => {
+    before(() => {
+      browser.url('/#/raw/tests/terra-date-picker/date-picker/date-picker-default');
+      browser.execute(() => {
+        // Removes the blinking cursor to prevent screenshot mismatches.
+        document.querySelector('input[name="terra-date-date-input"]').style.caretColor = 'transparent';
+      });
+    });
+
+    it('sets the date', () => {
+      browser.setValue('input[name="terra-date-date-input"]', '0');
+      browser.keys('a1.b2/;3');
+    });
+
+    Terra.should.validateElement('default', { axeRules: { rules: ignoredA11y } });
+  });
 });